requirements.txt
django==3.1.4
djangorestframework==3.12.4
psycopg2==2.8.6
alpha-vantage==2.3.1
celery==5.0.5
django-extensions==3.1.3